requirements.txt
asgiref==3.11.0
brotli==1.2.0
click==8.3.1
colorama==0.4.6
dj-database-url==3.0.1
Django==6.0
gunicorn==23.0.0
h11==0.16.0
packaging==25.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.4
tzdata==2025.3
uvicorn==0.40.0
whitenoise==6.11.0
reportlab==4.2.5
Pillow==10.1.0
